style(robustness): drop redundant int cast on math.ceil

+++ src/raitap/robustness/assessors/marabou_assessor.py
@@ -252,7 +252,7 @@ class MarabouAssessor(FormalVerificationAssessor):
             network.addDisjunctionConstraint(disjuncts)
 
         options = Marabou.createOptions(
-            timeoutInSeconds=max(1, int(math.ceil(self.timeout_s))), verbosity=0
+            timeoutInSeconds=max(1, math.ceil(self.timeout_s)), verbosity=0
         )
         started = time.perf_counter()
         exit_code, values, stats = network.solve(options=options)
@@ -416,9 +416,7 @@ def _bisect_output_bound(
     from maraboupy import Marabou  # type: ignore[import-not-found]
 
     lo, hi = -float(search_range), float(search_range)
-    options = Marabou.createOptions(
-        timeoutInSeconds=max(1, int(math.ceil(timeout_s))), verbosity=0
-    )
+    options = Marabou.createOptions(timeoutInSeconds=max(1, math.ceil(timeout_s)), verbosity=0)
     max_iters = max(1, math.ceil(math.log2((2.0 * search_range) / tolerance)) + 2)
     had_certifying_unsat = False
     for _ in range(max_iters):
